requirements.txt
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-appconf==1.0.3
django-ckeditor==5.7.0
django-js-asset==1.2.2
django-model-utils==3.1.2
django-modelcluster==4.4
django-redis==4.10.0
django-taggit==0.24.0
django-treebeard==4.3
django-users2==0.2.2
djangorestframework==3.9.3
draftjs-exporter==2.1.5
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
mysqlclient==1.4.2.post1
Pillow==5.4.1
PySnooper==0.0.38
pytz==2019.1
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.24.3
wagtail==2.5.1
webencodings==0.5.1
Willow==1.1
